patch from author for test failure on 32-bit macs
git-svn-id: svn://tug.org/texlive/trunk/Build/source@29582 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/chktex/chktex-1.7.1/FindErrs.c
@@ -926,7 +926,7 @@ static void CheckRest(void)
                         /* Default -- show the match */
                         PSERR2(offset + MATCH.rm_so, MATCH.rm_eo - MATCH.rm_so,
                                emUserWarnRegex,
-                               MATCH.rm_eo - MATCH.rm_so,
+                               (int)(MATCH.rm_eo - MATCH.rm_so),
                                TmpBuffer + offset + MATCH.rm_so);
                     }
                     offset += MATCH.rm_eo;
